feat: split the number field into a control and field composables
