+++ packages/core/src/useNumberField/index.ts
@@ -1,1 +1,2 @@
 export * from './useNumberField';
+export * from './useNumberControl';
