+++ packages/core/src/useNumberField/useNumberControl.ts
@@ -1,0 +1,301 @@
+import { computed, nextTick, shallowRef, toValue, watch } from 'vue';
+import {
+  fromNumberish,
+  isEmpty,
+  isNullOrUndefined,
+  normalizeProps,
+  propsToValues,
+  useUniqId,
+  useCaptureProps,
+} from '../utils/common';
+import {
+  AriaDescribableProps,
+  AriaLabelableProps,
+  InputEvents,
+  AriaValidatableProps,
+  Numberish,
+  Reactivify,
+} from '../types';
+import { useInputValidity } from '../validation/useInputValidity';
+import { useNumberParser } from '../i18n/useNumberParser';
+import { useSpinButton } from '../useSpinButton';
+import { useLocale } from '../i18n';
+import { FormField, useFormFieldContext } from '../useFormField';
+import { FieldTypePrefixes } from '../constants';
+import { useEventListener } from '../helpers/useEventListener';
+import { TransparentWrapper } from '../types';
+import { useVModelProxy } from '../reactivity/useVModelProxy';
+
+export interface NumberInputDOMAttributes {
+  name?: string;
+}
+
+export interface NumberInputDOMProps
+  extends NumberInputDOMAttributes,
+    AriaLabelableProps,
+    AriaDescribableProps,
+    AriaValidatableProps,
+    InputEvents {
+  id: string;
+}
+
+export interface NumberControlProps {
+  /**
+   * The locale to use for number formatting.
+   */
+  locale?: string;
+
+  /**
+   * The v-model value of the number field.
+   */
+  modelValue?: number;
+
+  /**
+   * The label text for the increment button.
+   */
+  incrementLabel?: string;
+
+  /**
+   * The label text for the decrement button.
+   */
+  decrementLabel?: string;
+
+  /**
+   * The name attribute for the number field input.
+   */
+  name?: string;
+
+  /**
+   * The value attribute of the number field input.
+   */
+  value?: Numberish;
+
+  /**
+   * The minimum allowed value.
+   */
+  min?: Numberish;
+
+  /**
+   * The maximum allowed value.
+   */
+  max?: Numberish;
+
+  /**
+   * The amount to increment/decrement by.
+   */
+  step?: Numberish;
+
+  /**
+   * Placeholder text shown when the number field is empty.
+   */
+  placeholder?: string | undefined;
+
+  /**
+   * Whether the number field is required.
+   */
+  required?: boolean;
+
+  /**
+   * Whether the number field is readonly.
+   */
+  readonly?: boolean;
+
+  /**
+   * Whether the number field is disabled.
+   */
+  disabled?: boolean;
+
+  /**
+   * Options for number formatting.
+   */
+  formatOptions?: Intl.NumberFormatOptions;
+
+  /**
+   * Whether to disable mouse wheel input.
+   */
+  disableWheel?: boolean;
+
+  /**
+   * Whether to disable HTML5 form validation.
+   */
+  disableHtmlValidation?: TransparentWrapper<boolean>;
+}
+
+export interface NumberControlExtras {
+  field?: FormField<number>;
+}
+
+export function useNumberControl(_props: Reactivify<NumberControlProps>, ctx?: NumberControlExtras) {
+  const props = normalizeProps(_props);
+  const inputId = useUniqId(FieldTypePrefixes.NumberField);
+  const inputEl = shallowRef<HTMLInputElement>();
+  const field = ctx?.field ?? useFormFieldContext();
+  const { locale } = useLocale(props.locale);
+  const parser = useNumberParser(locale, props.formatOptions);
+  const { model, setModelValue } = useVModelProxy(field);
+
+  const isDisabled = computed(() => toValue(props.disabled) || field?.isDisabled.value);
+  const formattedText = shallowRef<string>('');
+  let updateValidity: () => void;
+
+  if (field) {
+    updateValidity = useInputValidity({
+      inputEl,
+      field,
+      disableHtmlValidation: props.disableHtmlValidation,
+    }).updateValidity;
+
+    field.registerControl({
+      getControlElement: () => inputEl.value,
+      getControlId: () => inputId,
+    });
+  }
+
+  watch(
+    [locale, () => toValue(props.formatOptions), model],
+    () => {
+      if (Number.isNaN(model.value) || isEmpty(model.value)) {
+        formattedText.value = '';
+
+        return;
+      }
+
+      formattedText.value = parser.format(model.value);
+    },
+    { immediate: true },
+  );
+
+  const { incrementButtonProps, decrementButtonProps, increment, decrement, spinButtonProps, applyClamp } =
+    useSpinButton({
+      current: model,
+      currentText: formattedText,
+      step: props.step,
+      min: props.min,
+      max: props.max,
+      readonly: props.readonly,
+      disabled: () => isDisabled.value || toValue(props.readonly),
+      incrementLabel: props.incrementLabel,
+      decrementLabel: props.decrementLabel,
+      orientation: 'vertical',
+      preventTabIndex: true,
+
+      onChange: value => {
+        setModelValue(value);
+        field?.setTouched(true);
+        updateValidity();
+      },
+    });
+
+  const handlers: InputEvents = {
+    onBeforeinput: (event: Event) => {
+      const inputEvent = event as InputEvent;
+      // No data,like backspace or whatever
+      if (isNullOrUndefined(inputEvent.data)) {
+        return;
+      }
+
+      const el = inputEvent.target as HTMLInputElement;
+      // Kind of predicts the next value of the input by appending the new data
+      const nextValue =
+        el.value.slice(0, el.selectionStart ?? undefined) +
+        inputEvent.data +
+        el.value.slice(el.selectionEnd ?? undefined);
+
+      const isValid = parser.isValidNumberPart(nextValue);
+      if (!isValid) {
+        inputEvent.preventDefault();
+        inputEvent.stopPropagation();
+        return;
+      }
+    },
+    onChange: (event: Event) => {
+      setModelValue(applyClamp(parser.parse((event.target as HTMLInputElement).value)));
+      nextTick(() => {
+        if (inputEl.value && inputEl.value?.value !== formattedText.value) {
+          inputEl.value.value = formattedText.value;
+        }
+      });
+    },
+    onBlur: () => {
+      field?.setTouched(true);
+    },
+  };
+
+  const inputMode = computed(() => {
+    const intlOpts = toValue(props.formatOptions);
+    const step = fromNumberish(props.step) || 1;
+    const hasDecimals = (intlOpts?.maximumFractionDigits ?? 0) > 0 || String(step).includes('.');
+
+    if (hasDecimals) {
+      return 'decimal';
+    }
+
+    return 'numeric';
+  });
+
+  const inputProps = useCaptureProps<NumberInputDOMProps>(() => {
+    return {
+      ...propsToValues(props, ['name', 'placeholder', 'required', 'readonly']),
+      ...field?.labelledByProps.value,
+      ...field?.describedByProps.value,
+      ...field?.accessibleErrorProps.value,
+      ...handlers,
+      onKeydown: spinButtonProps.value.onKeydown,
+      id: inputId,
+      inputmode: inputMode.value,
+      value: formattedText.value,
+      disabled: isDisabled.value ? true : undefined,
+      max: toValue(props.max),
+      min: toValue(props.min),
+      type: 'text',
+      spellcheck: false,
+    };
+  }, inputEl);
+
+  useEventListener(
+    inputEl,
+    'wheel',
+    (e: WheelEvent) => {
+      if (e.deltaY > 0) {
+        increment();
+        return;
+      }
+
+      decrement();
+    },
+    { disabled: () => isDisabled.value || toValue(props.disableWheel), passive: true },
+  );
+
+  return {
+    /**
+     * Decrements the number field value.
+     */
+    decrement,
+    /**
+     * Props for the decrement button.
+     */
+    decrementButtonProps,
+
+    /**
+     * Increments the number field value.
+     */
+    increment,
+    /**
+     * Props for the increment button.
+     */
+    incrementButtonProps,
+    /**
+     * Reference to the input element.
+     */
+    inputEl,
+    /**
+     * Props for the input element.
+     */
+    inputProps,
+
+    /**
+     * The formatted text of the number field.
+     */
+    formattedText,
+  };
+}
